requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
arrow==1.2.1
    # via jinja2-time
binaryornot==0.4.4
    # via cookiecutter
certifi==2021.10.8
    # via requests
chardet==4.0.0
    # via binaryornot
charset-normalizer==2.0.10
    # via requests
click==8.0.3
    # via cookiecutter
cookiecutter==1.7.3
    # via -r requirements.in
idna==3.3
    # via requests
jinja2==3.0.3
    # via
    #   cookiecutter
    #   jinja2-time
jinja2-time==0.2.0
    # via cookiecutter
markupsafe==2.0.1
    # via jinja2
poyo==0.5.0
    # via cookiecutter
python-dateutil==2.8.2
    # via arrow
python-slugify==5.0.2
    # via cookiecutter
pyyaml==6.0
    # via -r requirements.in
requests==2.27.1
    # via cookiecutter
six==1.16.0
    # via
    #   cookiecutter
    #   python-dateutil
text-unidecode==1.3
    # via python-slugify
urllib3==1.26.7
    # via requests
